films from LocalStorage
добавил 2 функции для рендера фильмов с локал стореджа
