+++ src/html/header-library.html
@@ -106,8 +106,8 @@
       </nav>
     </div>
     <div class="header-library__buttons">
-      <button class="btn btn-active">Watched</button>
-      <button class="btn">queue</button>
+      <button class="btn btn-active" id="watched">Watched</button>
+      <button class="btn" id="queue">queue</button>
     </div>
   </div>
 </header>
